Updated to core 3.1 and file upload
Updated to ASP.NET COre 3.1.

File upload and handling is functional. It is possible to upload image files while editing a post, or upload using the Files admin menu. In the Files admin menu it is also possible to upload other files than images. These files can then be added as links to a page or post in the Add/Edit admin menu.